requirements.txt
python-dotenv
flask
flask_cors
flask_sqlalchemy
gunicorn
gevent
mysql-connector-python